requirements.txt
colorama==0.4.1
certifi>=2019.3.9
chardet==3.0.4
idna==2.8
paho-mqtt==1.4.0
requests==2.22.0
six==1.12.0
urllib3==1.25.3
